feat: update openapi definition to support the latest API

+++ _examples/jitoswap/main.go
@@ -16,7 +16,7 @@ func main() {
 
 	ctx := context.TODO()
 
-	slippageBps := 250
+	slippageBps := uint64(250)
 
 	// Get the current quote for a swap.
 	// Ensure that the input and output mints are valid.
@@ -41,13 +41,13 @@ func main() {
 
 	// Define the prioritization fee in lamports.
 	prioritizationFeeLamports := &struct {
-		JitoTipLamports              *int `json:"jitoTipLamports,omitempty"`
+		JitoTipLamports              *uint64 `json:"jitoTipLamports,omitempty"`
 		PriorityLevelWithMaxLamports *struct {
-			MaxLamports   *int                                                                                   `json:"maxLamports,omitempty"`
+			MaxLamports   *uint64                                                                                `json:"maxLamports,omitempty"`
 			PriorityLevel *jupiter.SwapRequestPrioritizationFeeLamportsPriorityLevelWithMaxLamportsPriorityLevel `json:"priorityLevel,omitempty"`
 		} `json:"priorityLevelWithMaxLamports,omitempty"`
 	}{
-		JitoTipLamports: new(int),
+		JitoTipLamports: new(uint64),
 	}
 
 	*prioritizationFeeLamports.JitoTipLamports = 1000
